+++ sandbox/src/sandbox.rs
@@ -92,7 +92,8 @@ impl SandboxInner {
             while let Async::Ready(Some(network)) = self.network_requests_rx.poll()? {
                 match network {
                     NetworkRequest::SendMessage(peer, msg) => self.sent.push_back((peer, msg)),
-                    NetworkRequest::DisconnectWithPeer(_) | NetworkRequest::Shutdown => {}
+                    NetworkRequest::DisconnectWithPeer(_) |
+                    NetworkRequest::Shutdown => {}
                 }
             }
             Ok(())
@@ -252,12 +253,7 @@ impl Sandbox {
     pub fn recv<T: Message>(&self, msg: T) {
         self.check_unexpected_message();
         // TODO Think about addresses.
-        let dummy_addr = SocketAddr::from((
-            [
-                127, 0, 0, 1
-            ],
-            12_039,
-        ));
+        let dummy_addr = SocketAddr::from(([127, 0, 0, 1], 12_039));
         let event = NetworkEvent::MessageReceived(dummy_addr, msg.raw().clone());
         self.inner.borrow_mut().handle_event(event);
     }
@@ -485,10 +481,9 @@ impl Sandbox {
 
     pub fn transactions_hashes(&self) -> Vec<Hash> {
         let node_state = self.node_state();
-        let rlock = node_state
-            .transactions()
-            .read()
-            .expect("Expected read lock");
+        let rlock = node_state.transactions().read().expect(
+            "Expected read lock",
+        );
         rlock.keys().cloned().collect()
     }
 
@@ -552,12 +547,16 @@ fn gen_primitive_socket_addr(idx: u8) -> SocketAddr {
 
 pub fn sandbox_with_services(services: Vec<Box<Service>>) -> Sandbox {
     let validators = vec![
-        gen_keypair_from_seed(&Seed::new([12; 32])), gen_keypair_from_seed(&Seed::new([13; 32])),
-        gen_keypair_from_seed(&Seed::new([16; 32])), gen_keypair_from_seed(&Seed::new([19; 32])),
+        gen_keypair_from_seed(&Seed::new([12; 32])),
+        gen_keypair_from_seed(&Seed::new([13; 32])),
+        gen_keypair_from_seed(&Seed::new([16; 32])),
+        gen_keypair_from_seed(&Seed::new([19; 32])),
     ];
     let service_keys = vec![
-        gen_keypair_from_seed(&Seed::new([20; 32])), gen_keypair_from_seed(&Seed::new([21; 32])),
-        gen_keypair_from_seed(&Seed::new([22; 32])), gen_keypair_from_seed(&Seed::new([23; 32])),
+        gen_keypair_from_seed(&Seed::new([20; 32])),
+        gen_keypair_from_seed(&Seed::new([21; 32])),
+        gen_keypair_from_seed(&Seed::new([22; 32])),
+        gen_keypair_from_seed(&Seed::new([23; 32])),
     ];
 
     let addresses: Vec<SocketAddr> = (1..5).map(gen_primitive_socket_addr).collect::<Vec<_>>();
@@ -650,7 +649,8 @@ pub fn sandbox_with_services(services: Vec<Box<Service>>) -> Sandbox {
 
 pub fn timestamping_sandbox() -> Sandbox {
     sandbox_with_services(vec![
-        Box::new(TimestampingService::new()), Box::new(ConfigUpdateService::new())
+        Box::new(TimestampingService::new()),
+        Box::new(ConfigUpdateService::new()),
     ])
 }
 
@@ -660,7 +660,6 @@ mod tests {
     use exonum::messages::{FromRaw, RawTransaction};
     use exonum::encoding;
     use exonum::crypto::{gen_keypair_from_seed, Seed};
-    use exonum::node::TransactionSend;
     use exonum::storage::Fork;
 
     use sandbox_tests_helper::{add_one_height, SandboxState, VALIDATOR_1, VALIDATOR_2,
@@ -720,7 +719,7 @@ mod tests {
 
         fn handle_commit(&self, context: &ServiceContext) {
             let tx = TxAfterCommit::new_with_height(context.height());
-            context.api_sender().send(Box::new(tx)).unwrap();
+            context.transaction_sender().send(Box::new(tx)).unwrap();
         }
     }
 
@@ -819,7 +818,8 @@ mod tests {
     #[test]
     fn test_sandbox_service_handle_commit() {
         let sandbox = sandbox_with_services(vec![
-            Box::new(HandleCommitService), Box::new(TimestampingService::new())
+            Box::new(HandleCommitService),
+            Box::new(TimestampingService::new()),
         ]);
         let state = SandboxState::new();
         add_one_height(&sandbox, &state);